+++ config.py
@@ -1,0 +1,15 @@
+# Authentication and channel configuration
+token = "your_token_here"
+channel_name = "your_channel_name_here"
+
+# Screen configuration
+length = 640
+height = 360
+fps = 60
+censor_radius = 15
+
+# PyAudio configuration
+volume = 0.5
+sample_rate = 44100
+num_channels = 1
+audio_format = pyaudio.paFloat32
